Add files via upload
This script takes a file (designed to be a proxy log CSV from Synology) and creates a new file, minus the lines containing text you put in the list (like 403 errors). This was built because pihole causes a ton of log flood when telemetry fails to connect to its remote end.